Functionify the post handler.
