+++ functions/invoice_functions.php
@@ -1,0 +1,519 @@
+<?php
+
+// Invoice related functions
+
+function createInvoice(
+    $client,
+    $scope,
+    $date,
+    $category
+) {
+
+    // Access global variables
+    global $mysqli, $session_company_currency, $session_user_id, $session_ip, $session_user_agent, $config_invoice_prefix, $config_invoice_next_number;
+
+    //Get Net Terms
+    $sql = mysqli_query($mysqli,"SELECT client_net_terms FROM clients WHERE client_id = $client");
+    $row = mysqli_fetch_array($sql);
+    $client_net_terms = intval($row['client_net_terms']);
+
+    //Get the last Invoice Number and add 1 for the new invoice number
+    $invoice_number = $config_invoice_next_number;
+    $new_config_invoice_next_number = $config_invoice_next_number + 1;
+    mysqli_query($mysqli,"UPDATE settings SET config_invoice_next_number = $new_config_invoice_next_number WHERE company_id = 1");
+
+    //Generate a unique URL key for clients to access
+    $url_key = randomString(156);
+
+    mysqli_query($mysqli,"INSERT INTO invoices SET invoice_prefix = '$config_invoice_prefix', invoice_number = $invoice_number, invoice_scope = '$scope', invoice_date = '$date', invoice_due = DATE_ADD('$date', INTERVAL $client_net_terms day), invoice_currency_code = '$session_company_currency', invoice_category_id = $category, invoice_status = 'Draft', invoice_url_key = '$url_key', invoice_client_id = $client");
+    $invoice_id = mysqli_insert_id($mysqli);
+
+    mysqli_query($mysqli,"INSERT INTO history SET history_status = 'Draft', history_description = 'INVOICE added!', history_invoice_id = $invoice_id");
+
+    //Logging
+    mysqli_query($mysqli,"INSERT INTO logs SET log_type = 'Invoice', log_action = 'Create', log_description = '$config_invoice_prefix$invoice_number', log_ip = '$session_ip', log_user_agent = '$session_user_agent', log_user_id = $session_user_id");
+
+}
+
+function readInvoice(
+    $invoice_id
+) {
+
+    // Access global variables
+    global $mysqli;
+
+    $sql = mysqli_query($mysqli,"SELECT * FROM invoices WHERE invoice_id = $invoice_id");
+    $row = mysqli_fetch_array($sql);
+
+    return $row;
+
+}
+
+function updateInvoice(
+    $invoice_id,
+    $invoice_due,
+    $invoice_status,
+    $invoice_prefix,
+    $invoice_number,
+    $invoice_scope,
+    $invoice_date,
+    $invoice_currency_code,
+    $invoice_category_id,
+    $invoice_url_key,
+    $invoice_client_id
+) {
+
+    // Access global variables
+    global $mysqli, $session_user_id, $session_ip, $session_user_agent;
+
+    mysqli_query($mysqli,"UPDATE invoices SET invoice_prefix = '$invoice_prefix', invoice_number = $invoice_number, invoice_scope = '$invoice_scope', invoice_date = '$invoice_date', invoice_due = '$invoice_due', invoice_currency_code = '$invoice_currency_code', invoice_category_id = $invoice_category_id, invoice_status = '$invoice_status', invoice_url_key = '$invoice_url_key', invoice_client_id = $invoice_client_id WHERE invoice_id = $invoice_id");
+
+    mysqli_query($mysqli,"INSERT INTO history SET history_status = '$invoice_status', history_description = 'INVOICE updated!', history_invoice_id = $invoice_id");
+
+    //Logging
+    mysqli_query($mysqli,"INSERT INTO logs SET log_type = 'Invoice', log_action = 'Update', log_description = '$invoice_prefix$invoice_number', log_ip = '$session_ip', log_user_agent = '$session_user_agent', log_user_id = $session_user_id");
+
+}
+
+function deleteInvoice(
+    $invoice_id
+) {
+
+    // Access global variables
+    global $mysqli, $session_user_id, $session_ip, $session_user_agent;
+
+    mysqli_query($mysqli,"DELETE FROM invoices WHERE invoice_id = $invoice_id");
+
+    //Delete Items Associated with the Invoice
+    $sql = mysqli_query($mysqli,"SELECT * FROM invoice_items WHERE item_invoice_id = $invoice_id");
+    while($row = mysqli_fetch_array($sql)) {
+        $item_id = intval($row['item_id']);
+        mysqli_query($mysqli,"DELETE FROM invoice_items WHERE item_id = $item_id");
+    }
+
+    //Delete History Associated with the Invoice
+    $sql = mysqli_query($mysqli,"SELECT * FROM history WHERE history_invoice_id = $invoice_id");
+    while($row = mysqli_fetch_array($sql)) {
+        $history_id = intval($row['history_id']);
+        mysqli_query($mysqli,"DELETE FROM history WHERE history_id = $history_id");
+    }
+
+    //Delete Payments Associated with the Invoice
+    $sql = mysqli_query($mysqli,"SELECT * FROM payments WHERE payment_invoice_id = $invoice_id");
+    while($row = mysqli_fetch_array($sql)) {
+        $payment_id = intval($row['payment_id']);
+        mysqli_query($mysqli,"DELETE FROM payments WHERE payment_id = $payment_id");
+    }
+
+    //unlink tickets from invoice
+    mysqli_query($mysqli,"UPDATE tickets SET ticket_invoice_id = 0 WHERE ticket_invoice_id = $invoice_id");
+
+    //Logging
+    mysqli_query($mysqli,"INSERT INTO logs SET log_type = 'Invoice', log_action = 'Delete', log_description = '$invoice_id', log_ip = '$session_ip', log_user_agent = '$session_user_agent', log_user_id = $session_user_id");
+}
+
+function copyInvoice(
+    $invoice_id,
+    $date
+) {
+
+    // Access global variables
+    global $mysqli, $session_user_id, $session_ip, $session_user_agent, $config_invoice_prefix, $config_invoice_next_number;
+
+    //Get Net Terms
+    $sql = mysqli_query($mysqli,"SELECT client_net_terms FROM clients, invoices WHERE client_id = invoice_client_id AND invoice_id = $invoice_id");
+    $row = mysqli_fetch_array($sql);
+    $client_net_terms = intval($row['client_net_terms']);
+
+    $invoice_number = $config_invoice_next_number;
+    $new_config_invoice_next_number = $config_invoice_next_number + 1;
+    mysqli_query($mysqli,"UPDATE settings SET config_invoice_next_number = $new_config_invoice_next_number WHERE company_id = 1");
+
+    $sql = mysqli_query($mysqli,"SELECT * FROM invoices WHERE invoice_id = $invoice_id");
+    $row = mysqli_fetch_array($sql);
+    $invoice_scope = sanitizeInput($row['invoice_scope']);
+    $invoice_discount_amount = floatval($row['invoice_discount_amount']);
+    $invoice_amount = floatval($row['invoice_amount']);
+    $invoice_currency_code = sanitizeInput($row['invoice_currency_code']);
+    $invoice_note = sanitizeInput($row['invoice_note']);
+    $client_id = intval($row['invoice_client_id']);
+    $category_id = intval($row['invoice_category_id']);
+
+    //Generate a unique URL key for clients to access
+    $url_key = randomString(156);
+
+
+    mysqli_query($mysqli,"INSERT INTO invoices SET invoice_prefix = '$config_invoice_prefix', invoice_number = $invoice_number, invoice_scope = '$invoice_scope', invoice_date = '$date', invoice_due = DATE_ADD('$date', INTERVAL $client_net_terms day), invoice_category_id = $category_id, invoice_status = 'Draft', invoice_discount_amount = $invoice_discount_amount, invoice_amount = $invoice_amount, invoice_currency_code = '$invoice_currency_code', invoice_note = '$invoice_note', invoice_url_key = '$url_key', invoice_client_id = $client_id");
+
+    $new_invoice_id = mysqli_insert_id($mysqli);
+
+    mysqli_query($mysqli,"INSERT INTO history SET history_status = 'Draft', history_description = 'Copied INVOICE!', history_invoice_id = $new_invoice_id");
+
+    $sql_items = mysqli_query($mysqli,"SELECT * FROM invoice_items WHERE item_invoice_id = $invoice_id");
+    while($row = mysqli_fetch_array($sql_items)) {
+        $item_id = intval($row['item_id']);
+        $item_name = sanitizeInput($row['item_name']);
+        $item_description = sanitizeInput($row['item_description']);
+        $item_quantity = floatval($row['item_quantity']);
+        $item_price = floatval($row['item_price']);
+        $item_subtotal = floatval($row['item_subtotal']);
+        $item_tax = floatval($row['item_tax']);
+        $item_total = floatval($row['item_total']);
+        $item_order = intval($row['item_order']);
+        $tax_id = intval($row['item_tax_id']);
+
+        mysqli_query($mysqli,"INSERT INTO invoice_items SET item_name = '$item_name', item_description = '$item_description', item_quantity = $item_quantity, item_price = $item_price, item_subtotal = $item_subtotal, item_tax = $item_tax, item_total = $item_total, item_order = $item_order, item_tax_id = $tax_id, item_invoice_id = $new_invoice_id");
+    }
+
+    //Logging
+    mysqli_query($mysqli,"INSERT INTO logs SET log_type = 'Invoice', log_action = 'Create', log_description = 'Copied Invoice', log_ip = '$session_ip', log_user_agent = '$session_user_agent', log_user_id = $session_user_id");
+
+}
+
+function createRecurringInvoice(
+    $client,
+    $frequency,
+    $start_date,
+    $category,
+    $scope
+) {
+    // Access global variables
+    global $mysqli, $session_company_currency, $session_user_id, $session_ip, $session_user_agent, $config_recurring_prefix, $config_recurring_next_number;
+
+    //Get the last Recurring Number and add 1 for the new Recurring number
+    $recurring_number = $config_recurring_next_number;
+    $new_config_recurring_next_number = $config_recurring_next_number + 1;
+    mysqli_query($mysqli,"UPDATE settings SET config_recurring_next_number = $new_config_recurring_next_number WHERE company_id = 1");
+
+    mysqli_query($mysqli,"INSERT INTO recurring SET recurring_prefix = '$config_recurring_prefix', recurring_number = $recurring_number, recurring_scope = '$scope', recurring_frequency = '$frequency', recurring_next_date = '$start_date', recurring_category_id = $category, recurring_status = 1, recurring_currency_code = '$session_company_currency', recurring_client_id = $client");
+
+    $recurring_id = mysqli_insert_id($mysqli);
+
+    mysqli_query($mysqli,"INSERT INTO history SET history_status = 'Active', history_description = 'Recurring Invoice created!', history_recurring_id = $recurring_id");
+
+    //Logging
+    mysqli_query($mysqli,"INSERT INTO logs SET log_type = 'Recurring', log_action = 'Create', log_description = '$start_date - $category', log_ip = '$session_ip', log_user_agent = '$session_user_agent', log_user_id = $session_user_id");
+}
+
+function createInvoiceFromRecurring(
+    $invoice_id,
+    $recurring_frequency
+){
+    // Access global variables
+    global $mysqli, $session_user_id, $session_ip, $session_user_agent, $config_recurring_prefix, $config_recurring_next_number;
+
+    $sql = mysqli_query($mysqli,"SELECT * FROM invoices WHERE invoice_id = $invoice_id");
+    $row = mysqli_fetch_array($sql);
+    $invoice_date = sanitizeInput($row['invoice_date']);
+    $invoice_amount = floatval($row['invoice_amount']);
+    $invoice_currency_code = sanitizeInput($row['invoice_currency_code']);
+    $invoice_scope = sanitizeInput($row['invoice_scope']);
+    $invoice_note = sanitizeInput($row['invoice_note']); //SQL Escape in case notes have , them
+    $client_id = intval($row['invoice_client_id']);
+    $category_id = intval($row['invoice_category_id']);
+
+    //Get the last Recurring Number and add 1 for the new Recurring number
+    $recurring_number = $config_recurring_next_number;
+    $new_config_recurring_next_number = $config_recurring_next_number + 1;
+    mysqli_query($mysqli,"UPDATE settings SET config_recurring_next_number = $new_config_recurring_next_number WHERE company_id = 1");
+
+    mysqli_query($mysqli,"INSERT INTO recurring SET recurring_prefix = '$config_recurring_prefix', recurring_number = $recurring_number, recurring_scope = '$invoice_scope', recurring_frequency = '$recurring_frequency', recurring_next_date = DATE_ADD('$invoice_date', INTERVAL 1 $recurring_frequency), recurring_status = 1, recurring_amount = $invoice_amount, recurring_currency_code = '$invoice_currency_code', recurring_note = '$invoice_note', recurring_category_id = $category_id, recurring_client_id = $client_id");
+
+    $recurring_id = mysqli_insert_id($mysqli);
+
+    mysqli_query($mysqli,"INSERT INTO history SET history_status = 'Draft', history_description = 'Recurring Created from INVOICE!', history_recurring_id = $recurring_id");
+
+    $sql_items = mysqli_query($mysqli,"SELECT * FROM invoice_items WHERE item_invoice_id = $invoice_id");
+    while($row = mysqli_fetch_array($sql_items)) {
+        $item_id = intval($row['item_id']);
+        $item_name = sanitizeInput($row['item_name']);
+        $item_description = sanitizeInput($row['item_description']);
+        $item_quantity = floatval($row['item_quantity']);
+        $item_price = floatval($row['item_price']);
+        $item_subtotal = floatval($row['item_subtotal']);
+        $item_tax = floatval($row['item_tax']);
+        $item_total = floatval($row['item_total']);
+        $item_order = intval($row['item_order']);
+        $tax_id = intval($row['item_tax_id']);
+
+        mysqli_query($mysqli,"INSERT INTO invoice_items SET item_name = '$item_name', item_description = '$item_description', item_quantity = $item_quantity, item_price = $item_price, item_subtotal = $item_subtotal, item_tax = $item_tax, item_total = $item_total, item_order = $item_order, item_tax_id = $tax_id, item_recurring_id = $recurring_id");
+    }
+
+    //Logging
+    mysqli_query($mysqli,"INSERT INTO logs SET log_type = 'Invoice', log_action = 'Create', log_description = 'From recurring invoice', log_ip = '$session_ip', log_user_agent = '$session_user_agent', log_user_id = $session_user_id");
+
+
+}
+
+function readRecurringInvoice(
+    $recurring_id
+) {
+    // Access global variables
+    global $mysqli;
+
+    $sql = mysqli_query($mysqli,"SELECT * FROM recurring WHERE recurring_id = $recurring_id");
+    $row = mysqli_fetch_array($sql);
+
+    return $row;
+}
+
+function updateRecurringInvoice(
+    $recurring_id,
+    $frequency,
+    $next_date,
+    $category,
+    $scope,
+    $status,
+    $recurring_discount
+) {
+    // Access global variables
+    global $mysqli, $session_user_id, $session_ip, $session_user_agent;
+
+    //Calculate new total
+    $sql = mysqli_query($mysqli,"SELECT * FROM invoice_items WHERE item_recurring_id = $recurring_id");
+    $recurring_amount = 0;
+    while($row = mysqli_fetch_array($sql)) {
+        $item_total = floatval($row['item_total']);
+        $recurring_amount = $recurring_amount + $item_total;
+    }
+    $recurring_amount = $recurring_amount - $recurring_discount;
+
+    mysqli_query($mysqli,"UPDATE recurring SET recurring_scope = '$scope', recurring_frequency = '$frequency', recurring_next_date = '$next_date', recurring_category_id = $category, recurring_discount_amount = $recurring_discount, recurring_amount = $recurring_amount, recurring_status = $status WHERE recurring_id = $recurring_id");
+
+    mysqli_query($mysqli,"INSERT INTO history SET history_status = '$status', history_description = 'Recurring modified', history_recurring_id = $recurring_id");
+
+    //Logging
+    mysqli_query($mysqli,"INSERT INTO logs SET log_type = 'Recurring', log_action = 'Modify', log_description = '$recurring_id', log_ip = '$session_ip', log_user_agent = '$session_user_agent', log_user_id = $session_user_id");
+
+}
+
+function deleteRecurringInvoice(
+    $recurring_id
+) {
+    // Access global variables
+    global $mysqli, $session_user_id, $session_ip, $session_user_agent;
+
+    mysqli_query($mysqli,"DELETE FROM recurring WHERE recurring_id = $recurring_id");
+
+    //Delete Items Associated with the Recurring
+    $sql = mysqli_query($mysqli,"SELECT * FROM invoice_items WHERE item_recurring_id = $recurring_id");
+    while($row = mysqli_fetch_array($sql)) {
+        $item_id = intval($row['item_id']);
+        mysqli_query($mysqli,"DELETE FROM invoice_items WHERE item_id = $item_id");
+    }
+
+    //Delete History Associated with the Invoice
+    $sql = mysqli_query($mysqli,"SELECT * FROM history WHERE history_recurring_id = $recurring_id");
+    while($row = mysqli_fetch_array($sql)) {
+        $history_id = intval($row['history_id']);
+        mysqli_query($mysqli,"DELETE FROM history WHERE history_id = $history_id");
+    }
+
+    //Logging
+    mysqli_query($mysqli,
+    "INSERT INTO logs SET
+        log_type = 'Recurring',
+        log_action = 'Delete',
+        log_description = '$recurring_id',
+        log_ip = '$session_ip',
+        log_user_agent = '$session_user_agent',
+        log_user_id = $session_user_id
+    ");
+}
+
+function updateInvoiceStatus(
+    $invoice_id,
+    $status
+) {
+    $archived_query = "";
+    switch($status):
+        case "Draft":
+            $history_description = "Invoice set to Draft";
+            break;
+        case "Sent":
+            $history_description = "Invoice sent to client";
+            break;
+        case "Viewed":
+            $history_description = "Invoice viewed by client";
+            break;
+        case "Paid":
+            $history_description = "Invoice paid by client";
+            break;
+        case "Overdue":
+            $history_description = "Invoice is overdue";
+            break;
+        case "Cancelled":
+            $history_description = "Invoice cancelled";
+            break;
+        case "Archived":
+            $history_description = "Invoice archived";
+            $archived_query = ", invoice_archived_at = NOW()";
+            break;
+        default:
+            $history_description = "Invoice status change error!";
+            $status = "Draft";
+            break;
+    endswitch;
+
+    // Access global variables
+    global $mysqli, $session_user_id, $session_ip, $session_user_agent;
+
+    mysqli_query($mysqli,"UPDATE invoices SET invoice_status = '$status'$archived_query WHERE invoice_id = $invoice_id");
+
+    mysqli_query($mysqli,"INSERT INTO history SET history_status = '$status', history_description = '$history_description', history_invoice_id = $invoice_id");
+
+    //Logging
+    mysqli_query($mysqli,"INSERT INTO logs SET log_type = 'Invoice', log_action = 'Status', log_description = 'Invoice ID $invoice_id', log_ip = '$session_ip', log_user_agent = '$session_user_agent', log_user_id = $session_user_id");
+    
+}
+
+function forceRecurring(
+    $recurring_id
+) {
+    // Access global variables
+    global $mysqli, $session_user_id, $session_ip, $session_user_agent, $config_recurring_auto_send_invoice, $config_invoice_next_number, $config_invoice_prefix, $config_invoice_from_email, $config_invoice_from_name, $config_base_url, $session_name;
+
+
+    $sql_recurring = mysqli_query($mysqli,"SELECT * FROM recurring, clients WHERE client_id = recurring_client_id AND recurring_id = $recurring_id");
+
+    $row = mysqli_fetch_array($sql_recurring);
+    $recurring_id = intval($row['recurring_id']);
+    $recurring_scope = sanitizeInput($row['recurring_scope']);
+    $recurring_frequency = sanitizeInput($row['recurring_frequency']);
+    $recurring_discount_amount = floatval($row['recurring_discount_amount']);
+    $recurring_amount = floatval($row['recurring_amount']);
+    $recurring_currency_code = sanitizeInput($row['recurring_currency_code']);
+    $recurring_note = sanitizeInput($row['recurring_note']);
+    $category_id = intval($row['recurring_category_id']);
+    $client_id = intval($row['recurring_client_id']);
+    $client_net_terms = intval($row['client_net_terms']);
+
+    //Get the last Invoice Number and add 1 for the new invoice number
+    $new_invoice_number = $config_invoice_next_number;
+    $new_config_invoice_next_number = $config_invoice_next_number + 1;
+    mysqli_query($mysqli,"UPDATE settings SET config_invoice_next_number = $new_config_invoice_next_number WHERE company_id = 1");
+
+    //Generate a unique URL key for clients to access
+    $url_key = randomString(156);
+
+    mysqli_query($mysqli,"INSERT INTO invoices SET invoice_prefix = '$config_invoice_prefix', invoice_number = $new_invoice_number, invoice_scope = '$recurring_scope', invoice_date = CURDATE(), invoice_due = DATE_ADD(CURDATE(), INTERVAL $client_net_terms day), invoice_discount_amount = $recurring_discount_amount, invoice_amount = $recurring_amount, invoice_currency_code = '$recurring_currency_code', invoice_note = '$recurring_note', invoice_category_id = $category_id, invoice_status = 'Sent', invoice_url_key = '$url_key', invoice_client_id = $client_id");
+
+    $new_invoice_id = mysqli_insert_id($mysqli);
+
+    //Copy Items from original invoice to new invoice
+    $sql_invoice_items = mysqli_query($mysqli,"SELECT * FROM invoice_items WHERE item_recurring_id = $recurring_id ORDER BY item_id ASC");
+
+    while($row = mysqli_fetch_array($sql_invoice_items)) {
+        $item_id = intval($row['item_id']);
+        $item_name = sanitizeInput($row['item_name']);
+        $item_description = sanitizeInput($row['item_description']);
+        $item_quantity = floatval($row['item_quantity']);
+        $item_price = floatval($row['item_price']);
+        $item_subtotal = floatval($row['item_subtotal']);
+        $item_order = intval($row['item_order']);
+        $tax_id = intval($row['item_tax_id']);
+
+        //Recalculate Item Tax since Tax percents can change.
+        if ($tax_id > 0) {
+            $sql = mysqli_query($mysqli,"SELECT * FROM taxes WHERE tax_id = $tax_id");
+            $row = mysqli_fetch_array($sql);
+            $tax_percent = floatval($row['tax_percent']);
+            $item_tax_amount = $item_subtotal * $tax_percent / 100;
+        } else {
+            $item_tax_amount = 0;
+        }
+
+        $item_total = $item_subtotal + $item_tax_amount;
+
+        //Update Recurring Items with new tax
+        mysqli_query($mysqli,"UPDATE invoice_items SET item_tax = $item_tax_amount, item_total = $item_total, item_tax_id = $tax_id, item_order = $item_order WHERE item_id = $item_id");
+
+        mysqli_query($mysqli,"INSERT INTO invoice_items SET item_name = '$item_name', item_description = '$item_description', item_quantity = $item_quantity, item_price = $item_price, item_subtotal = $item_subtotal, item_tax = $item_tax_amount, item_total = $item_total, item_tax_id = $tax_id, item_invoice_id = $new_invoice_id");
+    }
+
+    mysqli_query($mysqli,"INSERT INTO history SET history_status = 'Sent', history_description = 'Invoice Generated from Recurring!', history_invoice_id = $new_invoice_id");
+
+    //Update Recurring Balances by tallying up recurring items also update recurring dates
+    $sql_recurring_total = mysqli_query($mysqli,"SELECT SUM(item_total) AS recurring_total FROM invoice_items WHERE item_recurring_id = $recurring_id");
+    $row = mysqli_fetch_array($sql_recurring_total);
+    $new_recurring_amount = floatval($row['recurring_total']) - $recurring_discount_amount;
+
+    mysqli_query($mysqli,"UPDATE recurring SET recurring_amount = $new_recurring_amount, recurring_last_sent = CURDATE(), recurring_next_date = DATE_ADD(CURDATE(), INTERVAL 1 $recurring_frequency) WHERE recurring_id = $recurring_id");
+
+    //Also update the newly created invoice with the new amounts
+    mysqli_query($mysqli,"UPDATE invoices SET invoice_amount = $new_recurring_amount WHERE invoice_id = $new_invoice_id");
+
+    if ($config_recurring_auto_send_invoice == 1) {
+        $sql = mysqli_query($mysqli,"SELECT * FROM invoices
+            LEFT JOIN clients ON invoice_client_id = client_id
+            LEFT JOIN contacts ON clients.client_id = contacts.contact_client_id AND contact_primary = 1
+            WHERE invoice_id = $new_invoice_id"
+        );
+        $row = mysqli_fetch_array($sql);
+
+        $invoice_prefix = sanitizeInput($row['invoice_prefix']);
+        $invoice_number = intval($row['invoice_number']);
+        $invoice_scope = sanitizeInput($row['invoice_scope']);
+        $invoice_date = sanitizeInput($row['invoice_date']);
+        $invoice_due = sanitizeInput($row['invoice_due']);
+        $invoice_amount = floatval($row['invoice_amount']);
+        $invoice_url_key = sanitizeInput($row['invoice_url_key']);
+        $client_id = intval($row['client_id']);
+        $contact_name = sanitizeInput($row['contact_name']);
+        $contact_email = sanitizeInput($row['contact_email']);
+
+        $sql = mysqli_query($mysqli,"SELECT * FROM companies WHERE company_id = 1");
+        $row = mysqli_fetch_array($sql);
+        $company_name = sanitizeInput($row['company_name']);
+        $company_phone = sanitizeInput(formatPhoneNumber($row['company_phone']));
+
+        // Sanitize Config Vars
+        $config_invoice_from_email = sanitizeInput($config_invoice_from_email);
+        $config_invoice_from_name = sanitizeInput($config_invoice_from_name);
+
+        // Email to client
+
+        $subject = "$company_name Invoice $invoice_prefix$invoice_number";
+        $body = "Hello $contact_name,<br><br>An invoice regarding \"$invoice_scope\" has been generated. Please view the details below.<br><br>Invoice: $invoice_prefix$invoice_number<br>Issue Date: $invoice_date<br>Total: $$invoice_amount<br>Due Date: $invoice_due<br><br><br>To view your invoice, please click <a href=\'https://$config_base_url/guest_view_invoice.php?invoice_id=$new_invoice_id&url_key=$invoice_url_key\'>here</a>.<br><br><br>--<br>$company_name - Billing<br>$company_phone";
+
+
+        $data = [
+            [
+                'from' => $config_invoice_from_email,
+                'from_name' => $config_invoice_from_name,
+                'recipient' => $contact_email,
+                'recipient_name' => $contact_name,
+                'subject' => $subject,
+                'body' => $body
+            ]
+        ];
+        $mail = addToMailQueue($mysqli, $data);
+
+        if ($mail === true) {
+            // Add send history
+            mysqli_query($mysqli,"INSERT INTO history SET history_status = 'Sent', history_description = 'Force Emailed Invoice!', history_invoice_id = $new_invoice_id");
+
+            // Update Invoice Status to Sent
+            mysqli_query($mysqli,"UPDATE invoices SET invoice_status = 'Sent', invoice_client_id = $client_id WHERE invoice_id = $new_invoice_id");
+
+        } else {
+            // Error reporting
+            mysqli_query($mysqli,"INSERT INTO notifications SET notification_type = 'Mail', notification = 'Failed to send email to $contact_email', notification_client_id = $client_id");
+            mysqli_query($mysqli,"INSERT INTO logs SET log_type = 'Mail', log_action = 'Error', log_description = 'Failed to send email to $contact_email regarding $subject. $mail', log_ip = '$session_ip', log_user_agent = '$session_user_agent', log_user_id = $session_user_id");
+        }
+
+    } //End Recurring Invoices Loop
+
+    //Logging
+    mysqli_query($mysqli,"INSERT INTO logs SET log_type = 'Invoice', log_action = 'Create', log_description = '$session_name forced recurring invoice into an invoice', log_ip = '$session_ip', log_user_agent = '$session_user_agent', log_client_id = $client_id, log_user_id = $session_user_id, log_entity_id = $new_invoice_id");
+}
+
+function addInvoiceToTicket(
+    $invoice_id,
+    $ticket_id
+) {
+    // Access global variables
+    global $mysqli, $session_user_id, $session_ip, $session_user_agent;
+
+    mysqli_query($mysqli,"UPDATE invoices SET invoice_ticket_id = $ticket_id WHERE invoice_id = $invoice_id");
+}
