+++ post/account.php
@@ -5,67 +5,65 @@
  */
 
 if (isset($_POST['add_account'])) {
+    // Check if the user is an accountant
+    validateAccountantRole();
 
+    // Sanitize the input
     $name = sanitizeInput($_POST['name']);
     $opening_balance = floatval($_POST['opening_balance']);
     $currency_code = sanitizeInput($_POST['currency_code']);
     $notes = sanitizeInput($_POST['notes']);
     $type = intval($_POST['type']);
 
-    mysqli_query($mysqli,"INSERT INTO accounts SET account_name = '$name', opening_balance = $opening_balance, account_currency_code = '$currency_code', account_type ='$type', account_notes = '$notes'");
+    // Create the account
+    createAccount($name, $opening_balance, $currency_code, $notes, $type);
 
-    //Logging
-    mysqli_query($mysqli,"INSERT INTO logs SET log_type = 'Account', log_action = 'Create', log_description = '$name', log_ip = '$session_ip', log_user_agent = '$session_user_agent', log_user_id = $session_user_id");
-
-    $_SESSION['alert_message'] = "Account added";
-
-    header("Location: " . $_SERVER["HTTP_REFERER"]);
-
+    // Redirect to the accounts page with a success message
+    referWithAlert("Account created", "success", "accounts.php");
 }
 
 if (isset($_POST['edit_account'])) {
+    // Check if the user is an accountant
+    validateAccountantRole();
 
+    // Sanitize the input
     $account_id = intval($_POST['account_id']);
     $name = sanitizeInput($_POST['name']);
     $type = intval($_POST['type']);
     $notes = sanitizeInput($_POST['notes']);
 
-    mysqli_query($mysqli,"UPDATE accounts SET account_name = '$name',account_type = '$type', account_notes = '$notes' WHERE account_id = $account_id");
+    // Edit the account
+    editAccount($account_id, $name, $type, $notes);
 
-    //Logging
-    mysqli_query($mysqli,"INSERT INTO logs SET log_type = 'Account', log_action = 'Modify', log_description = '$name', log_ip = '$session_ip', log_user_agent = '$session_user_agent', log_user_id = $session_user_id");
-
-    $_SESSION['alert_message'] = "Account modified";
-
-    header("Location: " . $_SERVER["HTTP_REFERER"]);
-
+    // Redirect to the referring page with a success message
+    referWithAlert("Account edited", "success");
 }
 
 if (isset($_GET['archive_account'])) {
+    // Check if the user is an accountant
+    validateAccountantRole();
+
+    // Sanitize the input
     $account_id = intval($_GET['archive_account']);
 
-    mysqli_query($mysqli,"UPDATE accounts SET account_archived_at = NOW() WHERE account_id = $account_id");
+    // Archive the account
+    archiveAccount($account_id);
 
-    //logging
-    mysqli_query($mysqli,"INSERT INTO logs SET log_type = 'Account', log_action = 'Archive', log_description = '$account_id', log_ip = '$session_ip', log_user_agent = '$session_user_agent'");
-
-    $_SESSION['alert_message'] = "Account Archived";
-
-    header("Location: " . $_SERVER["HTTP_REFERER"]);
-
+    // Redirect to the accounts page with a success message
+    referWithAlert("Account archived", "success", "accounts.php");
 }
 
 if (isset($_GET['delete_account'])) {
+    // Check if the user is an accountant
+    validateAccountantRole();
+
+    // Sanitize the input
     $account_id = intval($_GET['delete_account']);
 
-    mysqli_query($mysqli,"DELETE FROM accounts WHERE account_id = $account_id");
+    // Delete the account
+    deleteAccount($account_id);
 
-    //Logging
-    mysqli_query($mysqli,"INSERT INTO logs SET log_type = 'Account', log_action = 'Delete', log_description = '$account_id', log_ip = '$session_ip', log_user_agent = '$session_user_agent', log_user_id = $session_user_id");
-
-    $_SESSION['alert_message'] = "Account deleted";
-
-    header("Location: " . $_SERVER["HTTP_REFERER"]);
-
+    // Redirect to the accounts page with a success message
+    referWithAlert("Account deleted", "success", "accounts.php");
 }
